requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.5.0
openai>=1.3.0
aiohttp>=3.9.0
redis>=5.0.1 
requests>=2.31.0
html2text>=2020.1.16
lxml>=4.9.3